+++ include/ReversingSpace/Storage/Core.hpp
@@ -18,6 +18,8 @@
 #ifndef REVERSINGSPACE_STORAGE_CORE_HPP
 #define REVERSINGSPACE_STORAGE_CORE_HPP
 
+#include <ReversingSpace/GameFileSystem/API.hpp>
+
 // std::uint8_t
 #include <cstdint>
 
@@ -36,22 +38,16 @@ namespace reversingspace {
 	namespace storage {
 
 		// Forward for `File`.
-		class File;
+		class REVSPACE_GAMEFILESYSTEM_API File;
 
 		/// Shared pointer type for `File`.
 		using FilePointer = std::shared_ptr<File>;
 
 		// Forward for `Object`.
-		class View;
+		class REVSPACE_GAMEFILESYSTEM_API View;
 
 		/// Shared pointer type for `Object`.
 		using ViewPointer = std::shared_ptr<View>;
-
-		// Forward for `Directory`.
-		class Directory;
-
-		/// Shared pointer type for `Directory`.
-		using DirectoryPointer = std::shared_ptr<Directory>;
 
 		/// Storage format (to keep it consistent) for offset(s).
 		using StorageOffset = std::int64_t;